+++ edsnlp/pipes/ner/disorders/connective_tissue_disease/connective_tissue_disease.py
@@ -75,10 +75,10 @@ class ConnectiveTissueDiseaseMatcher(DisorderMatcher):
 
     Authors and citation
     --------------------
-    The `eds.connective_tissue_disease` component was developed by AP-HP's Data Science
-    team with a team of medical experts. A paper describing in details the development
-    of those components is being drafted and will soon be available.
-    """
+    The `eds.connective_tissue_disease` component was developed by AP-HP's Data Science team with a
+    team of medical experts, following the insights of the algorithm proposed
+    by [@petitjean_2024].
+    """  # noqa: E501
 
     def __init__(
         self,