+++ edsnlp/pipes/ner/disorders/peripheral_vascular_disease/peripheral_vascular_disease.py
@@ -76,10 +76,10 @@ class PeripheralVascularDiseaseMatcher(DisorderMatcher):
     Authors and citation
     --------------------
 
-    The `eds.peripheral_vascular_disease` component was developed by AP-HP's Data
-    Science team with a team of medical experts. A paper describing in details the
-    development of those components is being drafted and will soon be available.
-    """
+    The `eds.peripheral_vascular_disease` component was developed by AP-HP's Data Science team with a
+    team of medical experts, following the insights of the algorithm proposed
+    by [@petitjean_2024].
+    """  # noqa: E501
 
     def __init__(
         self,